requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic-settings>=2.2.1
sqlalchemy>=2.0.29
